Integer columns and better debug infra
The original intent of this commit was to allow integer columns like:

```
$ open Cargo.toml | get features.sys
- | -------
0 | heim
1 | battery
$ open Cargo.toml | get features.sys.0
heim
```

This change resulted in a number of infrastructure improvements.

First, this commit centralizes lookups around `ColumnPath`, a new primitive
value that represents a list of `PathMember`. A `PathMember` is either a string
or an integer.

`Tagged<Value>` now has `get_data_by_member` and `get_data_by_column_path`
methods, which produce a new `Tagged<Value>` by looking up a value at the mpath
member or column path respectively.

Relatedly, the insertion and updating methods were also changed to use member
and column path at their core.

NOTE: The dot operator is a direct translation into `get_data_by_member`, while
`| get` is subtly different. In particular, the dot operator looks up a member
on an object, while the `| get` command looks up a member for each row in a
table. It's important not to think of `ls | get a.b.c.d` as a direct
translation to `ls | get a | get b | get c | get d`. This is probably not
tenable in the long-term, but it will have to do for now.

Second, this commit substantially overhauls the debug infrastructure. This came
about because the anemic debug infrastructure made it too difficult to
implement the feature.

The important changes to the debug infrastructure are to standardize around
output built by `pretty.rs`.

The new `what?` command describes the type of inputs.

```
$ ls | first 2 | what?
0 | row accessed(date) created(date) modified(date) name(string) size(bytesize) type(string) |
1 | row accessed(date) created(date) modified(date) name(string) size(bytesize) type(string) |
```

The `debug` command is overhauled to be a lot clearer.

```
$ ls | first 2 | debug
0 | (name=".azure"
  |   type="Directory"
  |   size=0(bytesize)
  |   created=2019-06-22T21:08:59.555937900Z(date)
  |   accessed=2019-11-19T00:01:19.102766400Z(date)
  |   modified=2019-11-18T19:22:49.615613500Z(date))
1 | (name=".cargo"
  |   type="Directory"
  |   size=0(bytesize)
  |   created=2019-05-26T07:20:34.693053800Z(date)
  |   accessed=2019-11-19T00:01:19.102766400Z(date)
  |   modified=2019-11-04T16:02:31.165063200Z(date))
```

A new `ReturnSuccess`, `DebugValue` is used to specify that a value sent
through the pipeline should be presented in debug format.

Under the hood, a bunch of infrastructure build on top of `pretty.rs` makes it
easy to build documents that are printed in this format.

The `PrettyDebug` trait, as well as the `DebugDoc` and `DebugDocBuilder`
structs in `traits.rs` are the bulk of the sugar. In addition to sugar on top
of `pretty.rs`, the `DebugDoc` layer provides special document types like
`primitive`, `keyword`, `description`, etc. These document types are
`pretty::Doc`s under the hood, with an annotation layering designed for nu.
That annotation layering provides a consistent coloring output.

The pretty printing of values inside the table is also now handled using
the `pretty.rs` infrastructure. One nice property of this is that we
don't have to worry too much about exact layout in each piece of the
debug printing, and can defer the formatting (both width and coloring)
to the point where we're about to render the document into a piece of
the output for which we know the properties.